SIO-1131 add tests (participants, oi)

+++ oioioi/participants/views.py
@@ -1,6 +1,7 @@
 from django.shortcuts import get_object_or_404, redirect
 from django.core.exceptions import SuspiciousOperation
 from django.core.urlresolvers import reverse
+from django.http import HttpResponseForbidden
 from django.utils.translation import ugettext_lazy as _
 from django.template.response import TemplateResponse
 from oioioi.base.menu import account_menu_registry
@@ -29,9 +30,12 @@ def registration_view(request, contest_id):
     return rcontroller.registration_view(request)
 
 def unregistration_view(request, contest_id):
+    rcontroller = request.contest.controller.registration_controller()
     if request.method == 'POST':
         participant = get_object_or_404(Participant, contest=request.contest,
                 user=request.user)
+        if not rcontroller.can_edit_registration(request, participant):
+            return HttpResponseForbidden()
         participant.delete()
         return redirect('index')
     return TemplateResponse(request, 'participants/unregister.html')
